+++ src/test/cpp/tests1/test-is_mailaddress.cpp
@@ -14,7 +14,7 @@
 
 // テスト対象関数のヘッダーファイル
 //#include "util/string_ex.h" //依存関係が多いのでテスト対象の関数定義のみ抜き出し
-BOOL IsMailAddress(const wchar_t* pszBuf, int nBufLen, int* pnAddressLength);
+BOOL IsMailAddress(const wchar_t* pszBuf, size_t nBufLen, int* pnAddressLength);
 
 //////////////////////////////////////////////////////////////////////
 // テストマクロ
